finish

+++ azure-storage-cpp-lite/src/blob/blob_client_wrapper.cpp
@@ -791,7 +791,7 @@ namespace microsoft_azure {
             return;
         }
 
-        void blob_client_wrapper::download_blob_to_file(const std::string &container, const std::string &blob, const std::string &destPath, const uint64_t file_offset, const uint64_t size, size_t parallel)
+        void blob_client_wrapper::download_blob_to_file(const std::string &container, const std::string &blob, const std::string &destPath, const unsigned long long file_offset, const unsigned long long size, size_t parallel)
         {
             if(!is_valid())
             {
